Add licence
I was about to write an article that talks about some of the design choices in Kent Beck's example, and was looking for an existing C# port, so that I didn't have to type it all in from scratch. Googling, this repository is literally the first hit (in my [filter bubble](https://en.wikipedia.org/wiki/Filter_bubble), at least).

I'd like to fork it, but although it's on GitHub, it's legally not open for copying or forking without a licence. With a licence, the legality of copying and forking is covered by the licence. The MIT licence is the one I usually use for my own projects.